Add Khmer short intro. Update README.md
Languages are a chaotic mess. This is a design choice for now. 

Yes, really, you cynical know-it-all hipsters.

Forcing ourselves to make every change in 4-5 languages would add hours to the workload. 

So Flying Monkey Radio is broadcasting in Khmer, Indonesian, Russian and English. 

Maybe if we contact C, his people can assist. 

#MoveAtMediumSpeedAndBalanceThings